Uso do pacote Fluent para criar validações. Definição das validações em Email, Telefone, Cpf, NomeCompleto Implementação da validação de Email.cs

+++ IJ.Domain/Interfaces/ITelefoneRepository.cs
@@ -4,5 +4,5 @@ public interface ITelefoneRepository
 {
     public Guid IdTelefone { get; set; }
     public int Ddd { get; set; }
-    public int NumeroTelefone { get; set; }
+    public long NumeroTelefone { get; set; }
 }
